papers/making_the_shoe_fit/requirements.txt
jax[cpu]
flax
optax
tensorflow_datasets
tensorflow
wandb
tqdm
hydra-core